+++ SGE.Entidades/Zona.cs
@@ -13,8 +13,8 @@ namespace SGE.Entidades
 
         public int Id { get; set; }
         public string Nombre { get; set; }
-        public decimal Latitud { get; set; }
-        public decimal Longitud { get; set; }
+        public double Latitud { get; set; }
+        public double Longitud { get; set; }
         public decimal Radio { get; set; }
         public List<Transformador> Transformadores { get; set; }
